Implement PCA 3D -> 2D projection
I suspect this won't be the best projection, and I'll have to continue
with my plans to open an OpenGL window to adjust the camera to pick the
best projection.

Pull out no small amount of hair getting the geometries into a flat
(n, 3) array of points, and figuring out how to reconstruct the
geometries from the transformed (n, 2) array of points.

Closes #20