Decide that >2 nested geometries won't be handled
That is, multi-geometries contained within a GEOMETRYCOLLECTION will not
be flattened. This is because I spent an entire day playing with the
the lifetimes required to retrieve a ConstGeometry from a Geometry, and
a ConstGeometry from another ConstGeometry. Specifically, I couldn't
figure out how to _return_ a ConstGeometry that was created from another
ConstGeometry.