Add GeometryNoder utility for noding geometries
The act of "noding" a geometry is to find all intersections, and break
the geometry into a list of non-intersecting (except at the endpoints)
line segments.

GEOS provides many such Noders, and even a GeometryNoder to facilitate
noding a geometry (as the Noders operate on SegmentStrings), but the
GeometryNoder uses a hardcoded IteratedNoder.

I want to be able to pass an instance of a Noder to GeometryNoder to
use. This remains to be done.