Refactor graph creation and storage
Keep a POINT in the node, not just a Coordinate. This is to make working
with the produced graph easier, so that users (myself) don't have to
keep track of the original GeometryFactory and convert the Coordinates
to Points themselves.

Add helper methods and constructors intended to help build a graph from
the geom2graph TGF output.