Discover graph generation is broken in 3D
The geos Noder API will happily accept 3D geometries. However, it only
acts as if the geometries are 2D. This assumption is fairly baked in,
and adjusting the geos::geom::Coordinate object's methods to act on the
Z coordinate is insufficient.

The line-line intersection only calculates intersections in the XY
plane, with points differing only in the z-coordinate treated as
identical. This is enforced by their spatial index implementation (but
really, KD-trees are K DIMENSIONAL trees. It's laughable that their
implementation is 2D only) _and_ in the spatial-index-accelerated
intersection APIs.

I'm incredibly disappointed, in order of frustration, that
* This limitation wasn't documented
* This limitation isn't a part of the interface (it shouldn't accept 3D
  geometries, now should it?)
* This limitation exists