Add SFCGAL as dependency
I attempted to use submodules for CGAL and SFCGAL, but ran into
innumerable issues with ExternalProject, find_package, and
add_subdirectory.

Unfortunately, the easy solution for me (install the dependencies
system-wide) kinda makes assumptions that some beginner user probably
doesn't have the chops to figure out. I'd _really_ rather the process
for using this project be

    1. clone
    2. build
    3. use

This adds a pretty non-trivial step 1.5, "satisfy dependencies". On a
recent Ubuntu distribution, this is _easy_

    sudo apt install libsfcgal-dev libboost-serialization-dev

But I remember being VERY frustrated with online tutorials that told you
how to do things on Linux, without providing instructions for Windows.
I don't have a Windows box to test with, but I _believe_ that all of the
Python dependencies are simple to install on Windows, and that this is
the _first_ case in this project that sucks on Windows.

I'd really like to find a way to provide SFCGAL (and its transitive
dependencies GMP and Boost?) as a submodule, or at least pre-built
binaries? I'm not convinced that's possible (or even desirable?).

Really, if that was my goal, I should probably focus on using Python for
the geom2graph tool?

_sigh_