Add option to use flattened geometries to tools
Allows each tool to use the flattened format as input or output.
Make the --input-format and --output-format arguments consistent.

While the flattened format does shave off some time in certain
scenarios, it's not enough to make using the format worthwhile. The
biggest impact for I/O performance is the number of lines (likely
because newlines cause flushes).

Perhaps I need to find a way to read/write input and output in batches,
large enough to boost I/O performance, but small enough that running
through a complex pipeline with large numbers of geometries doesn't slam
from one tool to the next. Rather, input should trickle from one tool
through to the next in such a way that each tool essentially runs in
parallel (unless it has to soak up the input and operate on it all at
once like tools/project.py).

Closes #45.