Regenerate all examples with a script
This is part of what I want to do in issues #66 and #67. The next step
is to make assertions on the generated output, so that I can catch
regressions.

TODO: There are massive diffs with every run of the script. I suspect
there's something nondeterministic about the ordering of geometries?