Pin isort and black versions
Black is now stable!! But before becoming stable, they apparently
changed how exponents are formatted (and for the better!!) so pin the
versions, and reformat the things that need it.

Additionally, filter out the Python files in the geom2graph tool
submodules. While this doesn't affect the pipeline, as the submodules
aren't initialized, it _does_ impact how black and isort run in the
local developer environment (if someone were to copy-paste from the
workflow if they didn't have it setup in their IDE).