Fix failing test build
I think switching to a newer compiler since the last time I built this
changed the operator ADL rules concerning namespaces. Normally, to
lookup an operator from a namespace, the argument to the operator must
also be a member of the same namespace hierarchy.

Here though, all the arguments come from the geos namespace, so the
compiler was refusing to lookup the operators from generative::io.

Fixes #124