Fix per-vertex streamline creation
The CoordsIter, MapCoords, and MapCoordsInPlace traits from geo all
require the function used in the mapping be either Fn + Copy or FnOnce,
neither of which are suitable for a mapping that has a side-effect,
which streamline generation necessarily is.

So I defined my own MapCoordsInPlaceMut with blackjack and hookers.

The two downsides are:
1. It'd be better for geo to relax its closure constraints
2. The Polygon impl has to clone the polygon because the
   Polygon::exterior and Polygon::interiors fields are private.