Add $ORIGIN/lib RPATH
When you run the binaries with "cargo run", they discover the
liblog4cplus and libgeos DSOs just fine due to the
rustc-link-search=native= path configured in the build.rs

But if you run the raw binaries themselves, they can no longer find the
DSOs.

    $ LD_TRACE_LOADED_OBJECTS=1 ./target/release/geom2graph --help
        linux-vdso.so.1 (0x00007ffdc5f69000)
        liblog4cplus.so.0 => not found
        libgeos.so.3.12.2 => not found
        libstdc++.so.6 => /lib64/libstdc++.so.6 (0x00007f43ce200000)
        libgcc_s.so.1 => /lib64/libgcc_s.so.1 (0x00007f43ce49e000)
        libm.so.6 => /lib64/libm.so.6 (0x00007f43ce11f000)
        libc.so.6 => /lib64/libc.so.6 (0x00007f43cdf41000)
        /lib64/ld-linux-x86-64.so.2 (0x00007f43ce4de000)

So set the RPATH to look at ./lib/, relative to the binary.