wkt2svg: Use CSS transform for flipping Y-axis
According to [1], the transform attribute is not supported on the <svg>
element in all browsers. So use a CSS transform instead, which has
better support.

Unfortunately this still isn't supported by ImageMagick through
librsvg2, but it's at least more conformant with web standards. If you
install inkscape, then ImageMagick will use that instead of librsvg2 to
render the SVG, but it doesn't support the transform:scale CSS property,
and won't flip the image upside-down.

[1] https://developer.mozilla.org/en-US/docs/Web/SVG/Reference/Attribute/transform#browser_compatibility